+++ src/fidesops/ops/api/v1/endpoints/privacy_request_endpoints.py
@@ -1,4 +1,4 @@
-# pylint: disable=too-many-branches,too-many-locals,too-many-lines
+# pylint: disable=too-many-branches,too-many-locals,too-many-lines, too-many-statements
 
 import csv
 import io
@@ -60,8 +60,10 @@ from fidesops.ops.api.v1.urn_registry import (
 from fidesops.ops.common_exceptions import (
     EmailDispatchException,
     FunctionalityNotConfigured,
+    IdentityNotFoundException,
     IdentityVerificationException,
     NoCachedManualWebhookEntry,
+    PolicyNotFoundException,
     TraversalError,
     ValidationError,
 )
@@ -73,7 +75,7 @@ from fidesops.ops.models.connectionconfig import ConnectionConfig
 from fidesops.ops.models.datasetconfig import DatasetConfig
 from fidesops.ops.models.email import EmailConfig
 from fidesops.ops.models.manual_webhook import AccessManualWebhook
-from fidesops.ops.models.policy import CurrentStep, Policy, PolicyPreWebhook
+from fidesops.ops.models.policy import ActionType, CurrentStep, Policy, PolicyPreWebhook
 from fidesops.ops.models.privacy_request import (
     ExecutionLog,
     PrivacyRequest,
@@ -87,6 +89,7 @@ from fidesops.ops.schemas.dataset import (
 from fidesops.ops.schemas.email.email import (
     EmailActionType,
     FidesopsEmail,
+    RequestReceiptBodyParams,
     SubjectIdentityVerificationBodyParams,
 )
 from fidesops.ops.schemas.external_https import PrivacyRequestResumeFormat
@@ -103,7 +106,10 @@ from fidesops.ops.schemas.privacy_request import (
     RowCountRequest,
     VerificationCode,
 )
-from fidesops.ops.service.email.email_dispatch_service import dispatch_email_task
+from fidesops.ops.service.email.email_dispatch_service import (
+    dispatch_email,
+    dispatch_email_task,
+)
 from fidesops.ops.service.privacy_request.request_runner_service import (
     generate_id_verification_code,
     queue_privacy_request,
@@ -232,7 +238,10 @@ async def create_privacy_request(
                 )
                 created.append(privacy_request)
                 continue  # Skip further processing for this privacy request
-
+            if config.notifications.send_request_receipt_notification:
+                _send_privacy_request_receipt_email_to_user(
+                    db, policy, privacy_request_data.identity.email
+                )
             if not config.execution.require_manual_request_approval:
                 AuditLog.create(
                     db=db,
@@ -297,6 +306,40 @@ def _send_verification_code_to_user(
             "to_email": email,
         },
     )
+
+
+def _send_privacy_request_receipt_email_to_user(
+    db: Session, policy: Optional[Policy], email: Optional[str]
+) -> None:
+    """Helper function to send request receipt email to the user"""
+    if not email:
+        logger.error(
+            IdentityNotFoundException(
+                "Identity email was not found, so request receipt email could not be sent."
+            )
+        )
+        return
+    if not policy:
+        logger.error(
+            PolicyNotFoundException(
+                "Policy was not found, so request receipt email could not be sent."
+            )
+        )
+        return
+    request_types: Set[str] = set()
+    for action_type in ActionType:
+        if policy.get_rules_for_action(action_type=ActionType(action_type)):
+            request_types.add(action_type)
+    try:
+        dispatch_email(
+            db=db,
+            action_type=EmailActionType.PRIVACY_REQUEST_RECEIPT,
+            to_email=email,
+            email_body_params=RequestReceiptBodyParams(request_types=request_types),
+        )
+    except EmailDispatchException as exc:
+        # catch early since this failure isn't fatal to privacy request, unlike the subject id verification email
+        logger.info("Email dispatch failed with exception %s", exc)
 
 
 def privacy_request_csv_download(
@@ -1095,8 +1138,18 @@ async def verify_identification_code(
     )
     try:
         privacy_request.verify_identity(db, provided_code.code)
+        policy: Optional[Policy] = Policy.get(
+            db=db, object_id=privacy_request.policy_id
+        )
+        if config.notifications.send_request_receipt_notification:
+            _send_privacy_request_receipt_email_to_user(
+                db, policy, privacy_request.get_persisted_identity().email
+            )
     except IdentityVerificationException as exc:
         raise HTTPException(status_code=HTTP_400_BAD_REQUEST, detail=exc.message)
+    except EmailDispatchException as exc:
+        # not fatal to request lifecycle, do not raise error, continue with request
+        logger.info("Email dispatch failed with exception %s", exc)
     except PermissionError as exc:
         logger.info("Invalid verification code provided for %s.", privacy_request.id)
         raise HTTPException(status_code=HTTP_403_FORBIDDEN, detail=exc.args[0])