+++ src/fidesops/ops/core/config.py
@@ -154,6 +154,7 @@ class FidesopsNotificationSettings(FidesSettings):
     """Configuration settings for data subject and/or data processor notifications"""
 
     send_request_completion_notification: Optional[bool] = True
+    send_request_receipt_notification: Optional[bool] = True
 
     class Config:
         env_prefix = "FIDESOPS__NOTIFICATIONS__"