+++ src/fidesops/ops/email_templates/get_email_template.py
@@ -8,6 +8,7 @@ from fidesops.ops.email_templates.template_names import (
     EMAIL_ERASURE_REQUEST_FULFILLMENT,
     PRIVACY_REQUEST_COMPLETE_ACCESS_TEMPLATE,
     PRIVACY_REQUEST_COMPLETE_DELETION_TEMPLATE,
+    PRIVACY_REQUEST_RECEIPT_TEMPLATE,
     SUBJECT_IDENTITY_VERIFICATION_TEMPLATE,
 )
 from fidesops.ops.schemas.email.email import EmailActionType
@@ -27,6 +28,8 @@ def get_email_template(action_type: EmailActionType) -> Template:
         return template_env.get_template(SUBJECT_IDENTITY_VERIFICATION_TEMPLATE)
     if action_type == EmailActionType.EMAIL_ERASURE_REQUEST_FULFILLMENT:
         return template_env.get_template(EMAIL_ERASURE_REQUEST_FULFILLMENT)
+    if action_type == EmailActionType.PRIVACY_REQUEST_RECEIPT:
+        return template_env.get_template(PRIVACY_REQUEST_RECEIPT_TEMPLATE)
     if action_type == EmailActionType.PRIVACY_REQUEST_COMPLETE_ACCESS:
         return template_env.get_template(PRIVACY_REQUEST_COMPLETE_ACCESS_TEMPLATE)
     if action_type == EmailActionType.PRIVACY_REQUEST_COMPLETE_DELETION: