+++ src/fidesops/ops/schemas/email/email.py
@@ -21,6 +21,7 @@ class EmailActionType(str, Enum):
     # verify email upon acct creation
     SUBJECT_IDENTITY_VERIFICATION = "subject_identity_verification"
     EMAIL_ERASURE_REQUEST_FULFILLMENT = "email_erasure_fulfillment"
+    PRIVACY_REQUEST_RECEIPT = "privacy_request_receipt"
     PRIVACY_REQUEST_COMPLETE_ACCESS = "privacy_request_complete_access"
     PRIVACY_REQUEST_COMPLETE_DELETION = "privacy_request_complete_deletion"
 
@@ -56,6 +57,12 @@ class FidesopsEmail(
         SubjectIdentityVerificationBodyParams,
         List[CheckpointActionRequired],
     ]
+
+
+class RequestReceiptBodyParams(BaseModel):
+    """Body params required for privacy request receipt email template"""
+
+    request_types: List[str]
 
 
 class AccessRequestCompleteBodyParams(BaseModel):