+++ src/fidesops/ops/service/email/email_dispatch_service.py
@@ -16,6 +16,7 @@ from fidesops.ops.schemas.email.email import (
     EmailServiceSecrets,
     EmailServiceType,
     FidesopsEmail,
+    RequestReceiptBodyParams,
     SubjectIdentityVerificationBodyParams,
 )
 from fidesops.ops.tasks import DatabaseTask, celery_app
@@ -51,6 +52,7 @@ def dispatch_email(
         Union[
             AccessRequestCompleteBodyParams,
             SubjectIdentityVerificationBodyParams,
+            RequestReceiptBodyParams,
             List[CheckpointActionRequired],
         ]
     ] = None,
@@ -105,6 +107,12 @@ def _build_email(
             body=base_template.render(
                 {"dataset_collection_action_required": body_params}
             ),
+        )
+    if action_type == EmailActionType.PRIVACY_REQUEST_RECEIPT:
+        base_template = get_email_template(action_type)
+        return EmailForActionType(
+            subject="Your request has been received",
+            body=base_template.render({"request_types": body_params.request_types}),
         )
     if action_type == EmailActionType.PRIVACY_REQUEST_COMPLETE_ACCESS:
         base_template = get_email_template(action_type)
@@ -167,4 +175,4 @@ def _mailgun_dispatcher(
             )
     except Exception as e:
         logger.error("Email failed to send: %s", Pii(str(e)))
-        raise EmailDispatchException(f"Email failed to send due to: {e}")
+        raise EmailDispatchException(f"Email failed to send due to: {Pii(e)}")